Source refactor and all that.
Packet reflection, and caching = done. (Planning on rewriting in C++ in
the future for performance improvements)

Readable Duplex Stream (ByteBuf) done. Yet to implement a writable
stream for creating network buffers.

Started on Habbo Encryption, yet haven't had time to if the encrypyted
keys the server generates are valid to the Habbo client.

Will eventually rewrite the logging framework as my own, and ditch
winston.